@@ -1,12 +1,12 @@
 name: malison
-version: 0.11.4
+version: 0.12.0
 author: Bob Nystrom <robert@stuffwithstuff.com>
 description: A little terminal UI package for the browser.
 homepage: https://github.com/munificent/malison
 dependencies:
-  piecemeal: ">=0.3.7 <0.5.0"
+  piecemeal: "^1.0.0"
 dev_dependencies:
   build_runner: any
   build_web_compilers: any
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
